+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - **Sheet Changes**
   - Improved display of changes to signature variables in sheet transactions
   - Question styling now alternates dynamically based on hidden questions
+- **Gem Changes**
+  - Updated to Ruby 2.3.2
 
 ### Refactoring
 - Removed deprecated_grid_variables from variables